Refactor authentications JWT and create likes app with models, serializers, views
