+++ authentications/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
+from rest_framework_simplejwt.views import TokenObtainPairView
 from . import views
 
 urlpatterns = [
-    # ...
-    path('login/', views.LoginView.as_view(), name='login'),
-    # ...
+    path('authentications/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
 ]
