+++ core/urls.py
@@ -6,8 +6,11 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+
+    path('api/v1/', include('authentications.urls')),
     path('api/v1/', include('pins.urls')),
     path('api/v1/', include('authentications.urls')),
     path('api/v1/', include('comments.urls')),
     path('api/v1/', include('ideas.urls')),
+    
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
